# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: taboo · 禁制 — 密钥管理平台 API
  version: 0.7.0
  description: 开源自部署密钥管理平台。高价值数据 + 严格访问控制 + 全程留痕。M5：Secret Sync / Webhooks / OIDC SSO / 审计导出。
servers:
  - url: http://localhost:7100
    description: 本地开发
...
paths:
...
  /api/v1/projects/{pid}/folders:
    get:
      tags: [folders]
      summary: 平铺文件夹列表（前端按 parent_id 建树）
      operationId: listFolders
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Pid' }, { $ref: '#/components/parameters/Env' }]
      responses:
        '200': { description: 文件夹列表, content: { application/json: { schema: { type: object, properties: { folders: { type: array, items: { $ref: '#/components/schemas/Folder' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [folders]
      summary: 创建文件夹（自动补中间节点，幂等）
      operationId: createFolder
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Pid' }, { $ref: '#/components/parameters/Env' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties: { path: { type: string, example: /db/mysql/ } }
      responses:
        '201': { description: 已创建, content: { application/json: { schema: { type: object, properties: { id: { type: string }, path: { type: string } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      tags: [folders]
      summary: 删除空文件夹（无子目录且无密钥；根目录拒绝）
      operationId: deleteFolder
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Pid' }
        - { $ref: '#/components/parameters/Env' }
        - { name: path, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: 已删除 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Pid: { name: pid, in: path, required: true, schema: { type: string } }
    Env: { name: env, in: query, schema: { type: string, default: dev } }
  responses:
    BadRequest: { description: 参数错误, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: 未认证/凭证错误, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    NotFound: { description: 资源不存在, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string, example: BAD_CREDENTIALS }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Folder:
      type: object
      properties:
        id: { type: string }
        parent_id: { type: string, description: 根目录为空串 }
        name: { type: string }
        path: { type: string, example: /db/mysql/ }
...